hv: ioapic: clean up HV_DEBUG usage
remove the usage of HV_DEBUG in ioapic.c

Tracked-On: #861
Acked-by: Eddie Dong <eddie.dong@intel.com>
